render_requirements.txt
telethon==1.35.0
python-dotenv==1.0.1
aiohttp==3.9.5
PyYAML==6.0.1